Fix cpplint's readability/casting by removing all unnecessary casts

+++ src/util/extra_math.cpp
@@ -26,5 +26,5 @@
 
 int extra_math::ceil_div(int num, int den)
 {
-    return (int)((num + (den - 1L)) / den);
+    return (num + (den - 1L)) / den;
 }
